Fixed bug of page reloading when scrolling
Page was reloading on scroll when I added check in Post.js to see if data was fetching and not showing the PostSummary view. Weird bug that I don't get why, but ot was reloading the page. But when putting the check for fetching inside PostSummary, then "No Posts" is only shown when data is empty and no fetching, making the make not reload on scroll to get more data.