Fixed bug of pages Steem, Feed and Blog not showing loader
When navigating between the pages for Steem, Feed and Blog, a common container component is used, Posts. The page loading spinner was not showing between routes, and would show the current/previous page until the data was fetched.

Emptying the redux array posts prior to fetching forced the loader to display and make it obvious something was happening.