Added box content for Home page Recent Kurations, improving UX
The data from recent curations for communities is no longer shown as a table by default, but with boxed content which includes an image from each post at the time it was added. If a post's main image has changed since the post was added to a community, it won't be changed in the community.