Fixed bug with revoked auth tokens not forcing logout
When a token expired or was revoked, the user would still stay logged in due to the cookie still being there. I fixed the check on returning users by checking the token with SteemConnect and making sure the token is valid before authenticating. When a token isn't valid, the user won't be logged in anymore.