Fixed bug with youtube videos being duplicated 9 times
A change I made to the steemit html ready file at some point in the past was supposed to be more up to date and valid, replacing a predicated method that was giving a warning. But in doing so, I apparently didn't do it right. So for now, I changed things back to the way busy/steemit does it and no more duplication of youtube videos.