Improved InfiniteScroll separation of concerns
When the Infinite Scroll functionality was previously put into Redux, I had still kept too much in the Posts component. Now the Posts component has been trimmed of some aspects of the infinite scroll functionality for Steem content, making it simpler, with the complexity for handling the infinte scroll residing in the Redux action creator.