Fixed useless loading symbol when no more data to fetch
When scrolling for posts, the last fetch for data should have set the 'hasMore' to false, but didn't. Instead it would do another fetch and set it when the data was empty. Instead, the 'hasMore' should be set to false when the results are less than the limit for pagination.