Added Clear button to Write posts page
When editing a post, but decided to not and cancel it, or to instead write a new post, there needed to be a way to clear the form data to start anew, without refreshing the page. Adding a clear button allows the page to be cleaned and ready for new data input.