Fixed edited comment losing child comments
When updating an edited comment, the child comments were lost. This was due to simply copying the updated comment fetch from Steem over the previous comment. When there were replies within the comment, they would be lose because the original comment fetching iterates to fetch child comments as long as there are some.