Fixed no auth user before data fetch
On some pages that used the user name in order to get user related data, the returning user verification in Redux was being returned only after the data fetch was already sent. This made some data return as a not logged in user, when the user was. This was due to the asynchronous redux-thunk in the app.

This has been corrected by checking if the redux store has the authenticated user, and if not, go get the username from the cookie instead of waiting for the asynchronous authentication.