Fix no auth on Home page, and not logged in bug
The Home page also had the need for the user name. The previous fix is now applied to the Home page.

The fix had an issue with non logged in user, as they had no cookie, thus the page would crash beacuse I wasn't testing for that, which I now do.