Fixed bug of route change from editing post to new post
When editing a post and then going to write a new one, the post that was edited was staying in view, and the form was not appearing. Clicking on the Write a second time had the form appear and route change. I had to add a check for the reset flag in teh Write component to make sure the edit form fisplayed, rather than the redirect for the edited post.