Fixed bug of spinner stuck on community page with no posts
A bug was introduced with some previous code changes, where the community page with no current posts was not showing the basic layout of no posts, members, and the name of the group. I was checking for posts in order to display, which was a blunder, since sometims there are not posts.