Fixed the view not moving transfered community to joined section
With the new transfer of ownership functionality, the owned group was no longer owned, but stayed in the owned section. To get it to go into the proper seciton, the easy and most efficient thing was to just call the data anew and the propagate it back down the child components again.