Fixed some posts not showing resteemed names
On some posts, the resteemed users are not present in the data returned by the javascript rpc package. This is an issue with dsteem, which I was first using. Even after using lightrpc which busy uses, I saw posts with multiple resteemed names not having any names at all. I have to use the original steem-js package from steemit in order to get all the resteemed names to show up properly.