Fixed calls for Steem data using steem-js
With the change in package for getting Steem data, I had to add a switch to determin which function to use to get data from Steem. The other javascript packages were better constructed to be dynamic in fetching data, but too bad they don't work properly to get all the data.