Fixed spinner loading when less than 20 results on summary pages
When a blog page, or feed, had less than 20 results, the loading spinner would activate when scrolling was done, but no more results were being returned since there were none. This caused the spinner and fetching to loop because hasMore was still true. Adding a check for results being less than the limit and setting hasMore to false fixed the issue.