Added interval updates to 'time ago' date formats
Date-times for posts on Steem and KURE would display the relative time when the page was loaded. I wanted to have upated time when the page was already laoded. The moment js package being used didn't have an interval setting. I switched to use the react-moment package instead, which has an interval settings.