Added grid layout UI/UX enhancements to Steem content
The Steem content pages now have an enhanced UI/UX that allows for modified content layouts for your viewing pleasure. Instead of only a list layout, a grid layout is provided with the View icon. In addition, you can toggle the description of posts if you want to read them before clicking to open a post.