Added icon showing if a post is flagged
Knowing when a post has been flagged is a valuable visible aid to have when looking at posts.  Rather than gray-out and censor the content from visibility, I opted to add a red outlined flag symbol next to the payout to show that a post has been flagged for some amount.